fix(events): fix adding listeners from JSX attributes
Closes #662
